requirements.txt
streamlit
huggingface_hub
pillow
requests
python-dotenv
openai
bcrypt
streamlit-authenticator
